Feat/editor refactor smaller components (#58)
* refactor to smaller components

* split up controls

* fix the fonts
