+++ packages/client/tailwind.config.ts
@@ -10,14 +10,14 @@ export default {
     extend: {
       fontFamily: {
         sans: [
-          'Inter"',
+          "Inter",
           "ui-sans-serif",
           "system-ui",
           "sans-serif",
-          'Apple Color Emoji"',
-          'Segoe UI Emoji"',
-          'Segoe UI Symbol"',
-          'Noto Color Emoji"',
+          "Apple Color Emoji",
+          "Segoe UI Emoji",
+          "Segoe UI Symbol",
+          "Noto Color Emoji",
         ],
       },
       borderRadius: {
