+++ packages/editor/package.json
@@ -1,12 +1,16 @@
 {
   "name": "@data-river/editor",
   "version": "0.1.0",
-  "main": "src/index.tsx",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc -b",
-    "test": "jest"
+    "watch": "tsc -b -w",
+    "test": "jest",
+    "lint": "eslint --fix 'src/**/*.{js,jsx,ts,tsx}'"
   },
   "dependencies": {
+    "lucide-react": "^0.441.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-flow": "^1.0.3",
