+++ packages/editor/src/Editor/components/controls/ControlSeparator.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+interface ControlSeparatorProps {}
+
+const ControlSeparator: React.FC<ControlSeparatorProps> = () => {
+  return <div className="w-px h-6 bg-gray-300 mx-2"></div>;
+};
+
+export default ControlSeparator;
