+++ packages/editor/src/index.tsx
@@ -1,5 +1,3 @@
 import Editor from "./Editor";
 
-export { default as Editor } from "./Editor";
-
 export default Editor;
